+++ src/index.js
@@ -1,6 +1,6 @@
 /**
  * @file jquery.add-input-area
- * @version 4.10.1
+ * @version 4.11.0
  * @author Yuusaku Miyazaki <toumin.m7@gmail.com>
  * @license MIT
  */